# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Planner API documentation",
        "description": "API Documentation for Travel Planner",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/trip": {
            "get": {
                "tags": [
                    "Trip"
                ],
                "summary": "Get all trips for logged in user",
                "description": "Get all trips excluding deleted ones",
                "operationId": "get_trip",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trip"
                ],
                "summary": "Create trip",
                "description": "Create trip",
                "operationId": "create_trip",
                "parameters": [
                    {
                        "name": "trip_name",
                        "in": "query",
                        "description": "string | max:255",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "origin",
                        "in": "query",
                        "description": "string | max:100",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "date_format:Y-m-d | after:today",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "date_format:Y-m-d | after:start_date",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "cost",
                        "in": "query",
                        "description": "numeric",
                        "schema": {
                            "type": "decimal"
                        }
                    },
                    {
                        "name": "group_type",
                        "in": "query",
                        "description": "in:SOLO,COUPLE,FAMILY,FRIENDS",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "trip_type",
                        "in": "query",
                        "description": "in:WORK,LEISURE",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "properties": {
                                "users": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                },
                                "destination": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "location": {
                                                "type": "string"
                                            },
                                            "start_date": {
                                                "type": "string"
                                            },
                                            "end_date": {
                                                "type": "string"
                                            },
                                            "cost": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "trip_banner": {
                                        "description": "image | mimes:jpeg, png, jpg, gif, svg | max:2048",
                                        "type": "file",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trip/update": {
            "post": {
                "tags": [
                    "Trip"
                ],
                "summary": "Update trip",
                "description": "Update trip",
                "operationId": "update_trip",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "required | exists:trips",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "trip_name",
                        "in": "query",
                        "description": "required | string | max:255",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "origin",
                        "in": "query",
                        "description": "string | max:100",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "date_format:Y-m-d | after:today",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "date_format:Y-m-d | after:start_date",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "cost",
                        "in": "query",
                        "description": "numeric",
                        "schema": {
                            "type": "decimal"
                        }
                    },
                    {
                        "name": "group_type",
                        "in": "query",
                        "description": "in:SOLO,COUPLE,FAMILY,FRIENDS",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "trip_type",
                        "in": "query",
                        "description": "in:WORK,LEISURE",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "trip_banner": {
                                        "description": "image | mimes:jpeg, png, jpg, gif, svg | max:2048",
                                        "type": "file",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trip/delete": {
            "post": {
                "tags": [
                    "Trip"
                ],
                "summary": "Delete trip",
                "description": "Delete trip",
                "operationId": "delete_trip",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "required | exists:trips,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}